requirements.txt
newsapi-python==0.2.7
python-dotenv==1.0.0

# Database packages
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# LLM Integration
xai-sdk>=1.4.0  # xAI SDK for Grok with native tool support (web_search, x_search, etc.)
pydantic==2.10.0  # Data validation and settings
tenacity==8.5.0  # Retry logic with exponential backoff (compatible with Streamlit)

# API Framework (future)
# fastapi==0.104.1
# uvicorn==0.24.0

# Streamlit Dashboard
streamlit>=1.40.0  # Updated for compatibility with xai-sdk packaging requirements

